Modified the XML platform files (and the scripts) with a <config> clause that sets a couple of command-line arguments under the hood:
<!--  WARNING:  This <config></config> clause below
makes it so that NO COMPUTATION TIME is simulated. This is because
in this module, for pedagogic purposes, we don't want to muddy the
(simulation) waters with computational times. As a results, this
XML platform file may not be suitable for running other
simulations, unless you remove the <config></config> clause.
-->
<config>
<prop id="smpi/simulate-computation" value="0"></prop>
<prop id="smpi/running-power" value="1"></prop>
</config>